docs: refactor vocabularies and update Markdown documentation
