+++ docs/example_data_fdp.csv
@@ -1,5 +1,5 @@
 id;name;description;author_name;author_id;keywords;publisher_name;publisher_id;theme;contact_point;issued
 1;Gryffindor research project;Impact of muggle technical inventions on word's magic presense;Hermione Granger;https://harrypotter.fandom.com/wiki/Hermione_Granger;magic, technic, muggles;House of Gryffindor;https://harrypotter.fandom.com/wiki/Gryffindor;https://harrypotter.fandom.com/wiki/Muggle_Studies;hgranger@sendmeowl.uk;03-04-1992
 2;Slytherin research project;Comarative analysis of magic powers of muggle-born and blood wizards ;Draco Malfoy;https://harrypotter.fandom.com/wiki/Draco_Malfoy;half-blood, muggle-born, pure-blood;House of Slytherin;https://harrypotter.fandom.com/wiki/Slytherin;https://harrypotter.fandom.com/wiki/Muggle_Studies, https://harrypotter.fandom.com/wiki/Muggle-born, https://harrypotter.fandom.com/wiki/Half-blood, https://harrypotter.fandom.com/wiki/Pure-blood;dmalfoy@sendmeowl.uk;25-05-1992
-3;Ravenclaw research project;Linguistic analysis of spells;Luna Lovegood;https://harrypotter.fandom.com/wiki/Luna_Lovegood;linguastics spell;House of Ravenclaw;https://harrypotter.fandom.com/wiki/Ravenclaw;https://harrypotter.fandom.com/wiki/List_of_spells;llovegood@sendmeowl.uk;26-05-1992
+3;Ravenclaw research project;Linguistic analysis of spells;Luna Lovegood;https://harrypotter.fandom.com/wiki/Luna_Lovegood;linguistics, spells;House of Ravenclaw;https://harrypotter.fandom.com/wiki/Ravenclaw;https://harrypotter.fandom.com/wiki/List_of_spells;llovegood@sendmeowl.uk;26-05-1992
 4;Hufflepuff research project;Domestics elfs reproduction rate changes in the last 100 years;Dora Williams;https://harrypotter.fandom.com/wiki/Dora_Williams;domestic elf;House of Hufflepuff;https://harrypotter.fandom.com/wiki/Hufflepuff;https://harrypotter.fandom.com/wiki/House-elf;dwilliams@sendmeowl.uk;27-05-1992
