+++ sempyro/hri_dcat/__init__.py
@@ -2,7 +2,7 @@ from .hri_catalog import HRICatalog
 from .hri_dataset import HRIDataset
 from .hri_data_service import HRIDataService
 from .hri_distribution import HRIDistribution
-from .vocabularies import GeonovumLicences
+from .vocabularies import GeonovumLicences, DatasetTheme
 
 __all__ = (
     "HRICatalog",
@@ -10,5 +10,6 @@ __all__ = (
     "HRIDistribution",
     "HRIDataService",
     # Enum classes
+    "DatasetTheme",
     "GeonovumLicences"
 )
